label_studio_ml/examples/detectron_instance_seg/requirements.txt
torch
torchvision
label-studio-sdk
opencv-python
detectron2@git+https://github.com/facebookresearch/detectron2.git
pillow